refactor: Option 페이지 생성 및 폴더 구조 수정
