+++ client/src/pages/making/complete/CompleteOptionPage.tsx
@@ -1,0 +1,5 @@
+function CompleteOptionPage() {
+  return <div></div>;
+}
+
+export default CompleteOptionPage;
